немного отрефакторил AccountHelper и перенес все функции связанные с данными в DataHandler. Тем самым избавился от God Object
