+++ app/src/main/java/com/kawunus/glossario/accountHelper/AccountHelper.kt
@@ -1,25 +1,18 @@
 package com.kawunus.glossario.accountHelper
 
-import android.content.Intent
 import android.widget.Toast
-import androidx.core.content.edit
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.google.android.gms.tasks.Task
-import com.google.firebase.auth.AuthResult
 import com.google.firebase.auth.GoogleAuthProvider
-import com.google.firebase.database.getValue
-import com.kawunus.glossario.MainActivity
-import com.kawunus.glossario.ProfileKeys
 import com.kawunus.glossario.R
 import com.kawunus.glossario.RegisterActivity
 
 
 class AccountHelper(private val act: RegisterActivity) {
     private lateinit var signInClient: GoogleSignInClient
-
+    private val dataHandler = DataHandler()
     fun signUpWithEmail(
         email: String, password: String, nickname: String
     ) {
@@ -27,7 +20,7 @@ class AccountHelper(private val act: RegisterActivity) {
             act.mAuth.createUserWithEmailAndPassword(email, password)
                 .addOnCompleteListener { task ->
                     if (task.isSuccessful) {
-                        sendDataToFirebase(task, nickname, email, "")
+                        dataHandler.sendRegistrationDataToFirebase(task, nickname, email, "", act)
                     } else {
                         showErrorToast(act.getString(R.string.sign_up_error_firebase))
                     }
@@ -43,7 +36,7 @@ class AccountHelper(private val act: RegisterActivity) {
         if (email.isNotEmpty() && password.isNotEmpty()) {
             act.mAuth.signInWithEmailAndPassword(email, password).addOnCompleteListener { task ->
                 if (task.isSuccessful) {
-                    getData(task)
+                    dataHandler.getDataFromFirebase(task, act)
                 } else {
                     showErrorToast(act.getString(R.string.sign_in_error_firebase))
                 }
@@ -71,8 +64,8 @@ class AccountHelper(private val act: RegisterActivity) {
         val credential = GoogleAuthProvider.getCredential(token, null)
         act.mAuth.signInWithCredential(credential).addOnCompleteListener { task ->
             if (task.isSuccessful) {
-                sendDataToFirebase(
-                    task, account.displayName!!, account.email!!, account.photoUrl!!.toString()
+                dataHandler.sendRegistrationDataToFirebase(
+                    task, account.displayName!!, account.email!!, account.photoUrl!!.toString(), act
                 )
             } else {
                 showErrorToast(act.getString(R.string.sign_in_error_google))
@@ -98,81 +91,4 @@ class AccountHelper(private val act: RegisterActivity) {
         Toast.makeText(act, message, Toast.LENGTH_LONG).show()
     }
 
-
-    private fun sendDataToFirebase(
-        task: Task<AuthResult>, nickname: String, email: String, imageURL: String
-    ) {
-        val user = task.result?.user
-        user?.let {
-            val userId = it.uid
-            val userEmail = it.email
-
-            val userMap = mapOf(
-                "email" to userEmail, "nickname" to nickname, "profileImage" to imageURL
-            )
-            act.database.child("users").child(userId).get().addOnCompleteListener { dataTask ->
-                if (dataTask.isSuccessful) {
-                    if (!dataTask.result.exists()) {
-
-                        act.database.child("users").child(userId).setValue(userMap)
-                            .addOnCompleteListener { setTask ->
-                                if (setTask.isSuccessful) {
-                                    saveData(email, nickname, imageURL)
-                                } else {
-                                    showErrorToast(
-                                        act.getString(
-                                            R.string.sign_up_error_save,
-                                            setTask.exception?.message
-                                        ))
-                                }
-                            }
-                    } else {
-                        getData(task)
-                    }
-                } else {
-                    showErrorToast(
-                        act.getString(
-                            R.string.sign_up_error_check,
-                            dataTask.exception?.message
-                        ))
-                }
-            }
-        }
-
-    }
-
-
-    private fun saveData(email: String, nickname: String, imageURL: String) {
-        act.prefs.edit(commit = true) {
-            putString(ProfileKeys.USER_STATUS, ProfileKeys.UserStatus.REGISTER)
-            putString(ProfileKeys.USER_EMAIL, email)
-            putString(ProfileKeys.USER_NICKNAME, nickname)
-            putString(ProfileKeys.USER_IMAGE, imageURL)
-        }
-
-        val intent = Intent(act, MainActivity::class.java)
-        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
-        act.startActivity(intent)
-        act.finish()
-    }
-
-    private fun getData(task: Task<AuthResult>) {
-        val user = task.result.user
-        user?.let {
-            val userId = it.uid
-
-            act.database.child("users").child(userId).get().addOnCompleteListener { dataTask ->
-                if (dataTask.isSuccessful) {
-                    val userData = dataTask.result?.getValue<Map<String, String>>()
-                    val userEmail = userData!!["email"] as String
-                    val userNickname = userData["nickname"] as String
-                    val imageUrl = userData["profileImage"] as String
-
-                    saveData(userEmail, userNickname, imageUrl)
-                } else {
-                    showErrorToast(act.getString(R.string.sign_up_error_get))
-                }
-            }
-        }
-    }
 }
